Removed unused objects and reduced Resource to FileResource
