+++ dc-model/src/main/java/de/digitalcollections/model/api/identifiable/Identifiable.java
@@ -2,7 +2,6 @@ package de.digitalcollections.model.api.identifiable;
 
 import de.digitalcollections.model.api.identifiable.parts.LocalizedText;
 import de.digitalcollections.model.api.identifiable.parts.structuredcontent.LocalizedStructuredContent;
-import de.digitalcollections.model.api.identifiable.resource.IiifImage;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
@@ -16,8 +15,6 @@ public interface Identifiable {
 
   LocalDateTime getLastModified();
 
-  IiifImage getIiifImage();
-
   IdentifiableType getType();
 
   UUID getUuid();
@@ -29,8 +26,6 @@ public interface Identifiable {
   void setLabel(LocalizedText label);
 
   void setLastModified(LocalDateTime lastModified);
-
-  void setIiifImage(IiifImage thumbnail);
 
   void setType(IdentifiableType identifiableType);
