+++ dc-model/src/main/java/de/digitalcollections/model/api/identifiable/IdentifiableType.java
@@ -4,7 +4,7 @@ package de.digitalcollections.model.api.identifiable;
  * All {@link Identifiable} types cudami can handle
  */
 public enum IdentifiableType {
-  ENTITY, ENTITY_PART, RESOURCE, RESOURCE_PART;
+  ENTITY, ENTITY_PART, RESOURCE;
 
   @Override
   public String toString() {
