--- dc-model/src/main/java/de/digitalcollections/model/api/identifiable/resource/IiifImage.java
@@ -1,14 +1,0 @@
-package de.digitalcollections.model.api.identifiable.resource;
-
-import java.net.URL;
-
-public interface IiifImage extends Resource {
-
-  String getIiifImageServiceUri();
-
-  URL getUrl();
-
-  void setIiifImageServiceUri(String iiifImageServiceUri);
-
-  void setUrl(URL url);
-}
