--- dc-model/src/main/java/de/digitalcollections/model/api/identifiable/resource/Resource.java
@@ -1,10 +1,0 @@
-package de.digitalcollections.model.api.identifiable.resource;
-
-import de.digitalcollections.model.api.identifiable.Identifiable;
-
-public interface Resource extends Identifiable {
-
-  ResourceType getResourceType();
-
-  void setResourceType(ResourceType resourceType);
-}
