--- dc-model/src/main/java/de/digitalcollections/model/api/identifiable/resource/ResourceType.java
@@ -1,13 +1,0 @@
-package de.digitalcollections.model.api.identifiable.resource;
-
-/**
- * All {@link Resource} types cudami can handle
- */
-public enum ResourceType {
-  BINARY_CONTENT, FILE, IIIF_IMAGE, IIIF_MANIFEST;
-
-  @Override
-  public String toString() {
-    return name();
-  }
-}
