--- dc-model/src/main/java/de/digitalcollections/model/impl/identifiable/resource/BinaryContentImpl.java
@@ -1,48 +1,0 @@
-package de.digitalcollections.model.impl.identifiable.resource;
-
-import de.digitalcollections.model.api.identifiable.resource.BinaryContent;
-import de.digitalcollections.model.api.identifiable.resource.MimeType;
-import de.digitalcollections.model.api.identifiable.resource.ResourceType;
-import java.net.URL;
-
-public class BinaryContentImpl extends ResourceImpl implements BinaryContent {
-
-  private MimeType mimeType;
-  private URL previewUrl;
-  private long sizeInBytes;
-
-  public BinaryContentImpl() {
-    super();
-    this.resourceType = ResourceType.BINARY_CONTENT;
-  }
-
-  @Override
-  public long getSizeInBytes() {
-    return sizeInBytes;
-  }
-
-  @Override
-  public void setSizeInBytes(long sizeInBytes) {
-    this.sizeInBytes = sizeInBytes;
-  }
-
-  @Override
-  public URL getPreviewUrl() {
-    return previewUrl;
-  }
-
-  @Override
-  public void setPreviewUrl(URL previewUrl) {
-    this.previewUrl = previewUrl;
-  }
-
-  @Override
-  public MimeType getMimeType() {
-    return mimeType;
-  }
-
-  @Override
-  public void setMimeType(MimeType mimeType) {
-    this.mimeType = mimeType;
-  }
-}
